--story=1005084 --user=龙儒锋 Tagent管理功能迁移 https://www.tapd.cn/54247054/s/1182920

See merge request codedriver/codedriver-tenant!398

+++ src/main/java/codedriver/module/tenant/api/runner/RunnerDeleteApi.java
@@ -9,10 +9,11 @@ import codedriver.framework.auth.core.AuthAction;
 import codedriver.framework.auth.label.RUNNER_MODIFY;
 import codedriver.framework.common.constvalue.ApiParamType;
 import codedriver.framework.dao.mapper.runner.RunnerMapper;
+import codedriver.framework.exception.runner.RunnerIdNotFoundException;
+import codedriver.framework.exception.runner.RunnerIsUsedException;
 import codedriver.framework.restful.annotation.*;
 import codedriver.framework.restful.constvalue.OperationTypeEnum;
 import codedriver.framework.restful.core.privateapi.PrivateApiComponentBase;
-import codedriver.framework.exception.runner.RunnerIdNotFoundException;
 import com.alibaba.fastjson.JSONObject;
 import org.springframework.stereotype.Service;
 
@@ -28,7 +29,7 @@ public class RunnerDeleteApi extends PrivateApiComponentBase {
 
     @Override
     public String getName() {
-        return "runner 删除接口";
+        return "删除runner";
     }
 
     @Override
@@ -52,6 +53,9 @@ public class RunnerDeleteApi extends PrivateApiComponentBase {
         Long id = paramObj.getLong("id");
         if (runnerMapper.checkRunnerIdIsExist(id) == 0) {
             throw new RunnerIdNotFoundException(id);
+        }
+        if (runnerMapper.checkRunnerIsUsed(id) > 0) {
+            throw new RunnerIsUsedException();
         }
         runnerMapper.deleteRunnerById(id);
         return null;
